Refactor for latest esteemer compatibility
